fix: now connection can be restored after SSLError (details in full comment)
When we got SSLError violation in Channel.send_method channel can be returned to pool and try to reused (channel.is_open == True), but actually it's not opened. Any operation with this will trigger errors like `OSError: Socket closed`, because internal transport closed.

If this behavior mixed with infinity ensures (retry_over_time in kombu) you can get "deadlock": waiting when connection restored in loop, but it's never happened